添加MGConst文件

+++ MGWeibo/Classes/Other(其他)/View/MGTabBar.h
@@ -15,7 +15,7 @@
 
 @protocol MGTabBarDelegate <NSObject>
 @optional
-- (void)tabBar:(MGTabBar *)tabBar didSelectedButtonFrom:(int)from to:(int)to;
+- (void)tabBar:(MGTabBar *)tabBar didSelectedButtonFrom:(NSInteger)from to:(NSInteger)to;
 /// tabBar点击了加号按钮
 - (void)tabBarDidClickedPlusBtton:(MGTabBar *)tabBar;
